CLN: strip out and form tools/concat.py from tools/merge.py
