+++ pandas/__init__.py
@@ -42,10 +42,10 @@ from pandas.core.api import *
 from pandas.sparse.api import *
 from pandas.stats.api import *
 from pandas.tseries.api import *
-from pandas.io.api import *
 from pandas.computation.api import *
 
-from pandas.tools.merge import (merge, concat, ordered_merge,
+from pandas.tools.concat import concat
+from pandas.tools.merge import (merge, ordered_merge,
                                 merge_ordered, merge_asof)
 from pandas.tools.pivot import pivot_table, crosstab
 from pandas.tools.plotting import scatter_matrix, plot_params
@@ -53,6 +53,8 @@ from pandas.tools.tile import cut, qcut
 from pandas.tools.util import to_numeric
 from pandas.core.reshape import melt
 from pandas.util.print_versions import show_versions
+
+from pandas.io.api import *
 
 # define the testing framework
 import pandas.util.testing
