+++ pandas/core/reshape.py
@@ -1194,7 +1194,7 @@ def get_dummies(data, prefix=None, prefix_sep='_', dummy_na=False,
     --------
     Series.str.get_dummies
     """
-    from pandas.tools.merge import concat
+    from pandas.tools.concat import concat
     from itertools import cycle
 
     if isinstance(data, DataFrame):
