+++ pandas/io/gbq.py
@@ -10,9 +10,7 @@ import sys
 import numpy as np
 
 from distutils.version import StrictVersion
-from pandas import compat
-from pandas.core.api import DataFrame
-from pandas.tools.merge import concat
+from pandas import compat, DataFrame, concat
 from pandas.core.common import PandasError
 from pandas.compat import lzip, bytes_to_str
